requirements.txt
# Simple RFP Screener Dependencies

anthropic>=0.39.0
python-dotenv>=1.0.0
